Misc tweaks and fixes
* EmuFramework: Split fast/slow mode into two separate functions
* EmuFramework: Add support for storing video aspect ratio separately for landscape/portrait modes
* EmuFramework: Remove redundant "Full Screen" and 8:7 aspect ratio options since custom values provide the same functions
* EmuFramework: Add "Square Pixels" aspect ratio option
* EmuFramework: Fix hq2x shader on GPUs with lower precision float math like the Mali-450
* EmuFramework: Add float cases for scanValue()
* Update about strings

+++ EmuFramework/src/GUIOptionView.cc
@@ -54,7 +54,7 @@ GUIOptionView::GUIOptionView(ViewAttachParams attach, bool customMenu):
 		{"Custom Value", &defaultFace(),
 			[this](const Input::Event &e)
 			{
-				app().pushAndShowNewCollectValueInputView<double>(attachParams(), e, "Input 2.0 to 10.0", "",
+				app().pushAndShowNewCollectValueInputView<float>(attachParams(), e, "Input 2.0 to 10.0", "",
 					[this](EmuApp &app, auto val)
 					{
 						int scaledIntVal = val * 1000.0;
